First code for mechatrinics shelf
The code takes in a .txt file with names and coordinates and puts those into a dictionary. It prits the list of names and asks you what you are looking for. It returns a string with servo coordinates in the form of 'x,y,r\n', to be used on an arduino.